refactor: PlaceRepository에서 HTTP 로직을 PlaceDataSource로 분리
- PlaceDataSource 인터페이스에 searchPlaceByCoordinates 메서드 추가
- TmapPlaceDataSourceImpl 생성 (TMAP API + Android Geocoder 로직 이동)
- PlaceRepositoryImpl 리팩토링 (558줄 → 58줄, 90% 감소)
- RemoteModule에 PlaceDataSource 바인딩 추가
- Clean Architecture 준수: Repository는 Domain 로직만, DataSource는 Data 접근 로직만
